Updated misc_session_status.test to not fail if select does not fail
The test with memory restrictions randomly works or fails in buildbot
depending on server configurations. On my machine the original test
worked.
As the test was there to just check if the server crashes when run with
small memory configurations, I disabled testing if the query would fail
or not. The test still has its original purpose.

Reviewed-by: Sergei Golubchik <serg@mariadb.org>