MDEV-35326: Memory Leak in init_io_cache_ext upon SHUTDOWN
'result' from m_lex of an instruction cleared in 3 places (1) cleanup()
(2) send_eof() (3) destructor but in case of deleting procedure just after
error non of them called.
(1) called before instruction execution of an instruction
(2) called at the end of successful execution of an instruction or
conventional execution
(3) called unconditionally at very end of conventional execution

In conventional execution it freed by direct call of the destructor in
THD::end_statement(), so lets call it when we destruct procedure (all
other cases will be covered by (1) and (2).